requirements.txt
flask
flask-cors
nltk
requests
google-genai
python-dotenv